move rest of backend files to backend folder
