+++ src/web/backend/playlists.go
@@ -1,4 +1,4 @@
-package web
+package backend
 
 import (
 	"encoding/json"
